+++ mobile/android/android-components/components/browser/engine-gecko-beta/src/main/java/mozilla/components/browser/engine/gecko/permission/GeckoPermissionRequest.kt
@@ -111,6 +111,10 @@ sealed class GeckoPermissionRequest constructor(
             callback.grant(videos.firstOrNull(), audios.firstOrNull())
         }
 
+        override fun containsVideoAndAudioSources(): Boolean {
+            return videoSources.isNotEmpty() && audioSources.isNotEmpty()
+        }
+
         override fun reject() {
             callback.reject()
         }