+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/permission/PermissionRequest.kt
@@ -4,6 +4,9 @@
 
 package mozilla.components.concept.engine.permission
 
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
 /**
  * Represents a permission request, used when engines need access to protected
  * resources. Every request must be handled by either calling [grant] or [reject].
@@ -47,6 +50,8 @@ interface PermissionRequest {
      * Rejects the requested permissions.
      */
     fun reject()
+
+    fun containsVideoAndAudioSources() = false
 }
 
 /**
@@ -55,26 +60,46 @@ interface PermissionRequest {
  * @property id an optional native engine-specific ID of this permission.
  * @property desc an optional description of what this permission type is for.
  */
-sealed class Permission(open val id: String? = "", open val desc: String? = "") {
+
+sealed class Permission(open val id: String? = "", open val desc: String? = "") : Parcelable {
+    @Parcelize
     data class ContentAudioCapture(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentAudioMicrophone(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentAudioOther(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentAutoplayMedia(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentGeoLocation(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentNotification(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentProtectedMediaId(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentVideoApplication(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentVideoBrowser(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentVideoCamera(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentVideoCapture(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentVideoScreen(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentVideoWindow(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class ContentVideoOther(override val id: String? = "", override val desc: String? = "") : Permission(id)
 
+    @Parcelize
     data class AppCamera(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class AppAudio(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class AppLocationCoarse(override val id: String? = "", override val desc: String? = "") : Permission(id)
+    @Parcelize
     data class AppLocationFine(override val id: String? = "", override val desc: String? = "") : Permission(id)
 
+    @Parcelize
     data class Generic(override val id: String?, override val desc: String? = "") : Permission(id)
 }